Move the state of AlbumDetails to App and pass the state and functions as props to albumDetails
